+++ docs/docs/integrations/embedding-models/voyage-ai.md
@@ -1,0 +1,25 @@
+---
+sidebar_position: 4
+---
+
+# Voyage AI
+
+## Maven Dependency
+
+```xml
+<dependency>
+    <groupId>dev.langchain4j</groupId>
+    <artifactId>langchain4j-voyage-ai</artifactId>
+    <version>0.35.0</version>
+</dependency>
+```
+
+## APIs
+
+- `VoyageEmbeddingModel`
+
+
+## Examples
+
+- [VoyageEmbeddingModelIT](https://github.com/langchain4j/langchain4j/blob/main/langchain4j-voyage/src/test/java/dev/langchain4j/model/voyage/VoyageEmbeddingModelIT.java)
+- [VoyageEmbeddingModelExample](https://github.com/langchain4j/langchain4j-examples/blob/main/voyage-examples/src/main/java/VoyageEmbeddingModelExample.java)
