+++ docs/docs/integrations/scoring-reranking-models/3-voyage-ai.md
@@ -1,0 +1,25 @@
+---
+sidebar_position: 2
+---
+
+# Voyage AI
+
+## Maven Dependency
+
+```xml
+<dependency>
+    <groupId>dev.langchain4j</groupId>
+    <artifactId>langchain4j-voyage-ai</artifactId>
+    <version>0.35.0</version>
+</dependency>
+```
+
+
+## Voyage Models
+
+- `VoyageScoringModel`
+
+
+## Examples
+- [Advanced RAG with re-ranking](https://github.com/langchain4j/langchain4j-examples/blob/main/rag-examples/src/main/java/_3_advanced/_03_Advanced_RAG_with_ReRanking_Example.java)
+- [VoyageScoringModelExample](https://github.com/langchain4j/langchain4j-examples/blob/main/voyage-examples/src/main/java/VoyageScoringModelExample.java)
