+++ langchain4j-voyage-ai/src/main/java/dev/langchain4j/model/voyageai/TokenUsage.java
@@ -1,0 +1,10 @@
+package dev.langchain4j.model.voyageai;
+
+class TokenUsage {
+
+    private Integer totalTokens;
+
+    public Integer getTotalTokens() {
+        return totalTokens;
+    }
+}
