--- langchain4j-voyage/src/main/java/dev/langchain4j/model/voyage/TokenUsage.java
@@ -1,10 +1,0 @@
-package dev.langchain4j.model.voyage;
-
-class TokenUsage {
-
-    private Integer totalTokens;
-
-    public Integer getTotalTokens() {
-        return totalTokens;
-    }
-}
