--- langchain4j-voyage/src/main/java/dev/langchain4j/model/voyage/VoyageApi.java
@@ -1,19 +1,0 @@
-package dev.langchain4j.model.voyage;
-
-import retrofit2.Call;
-import retrofit2.http.Body;
-import retrofit2.http.Headers;
-import retrofit2.http.POST;
-
-interface VoyageApi {
-
-    String DEFAULT_BASE_URL = "https://api.voyageai.com/v1/";
-
-    @POST("embeddings")
-    @Headers({"accept: application/json", "content-type: application/json"})
-    Call<EmbeddingResponse> embed(@Body EmbeddingRequest request);
-
-    @POST("rerank")
-    @Headers({"accept: application/json", "content-type: application/json"})
-    Call<RerankResponse> rerank(@Body RerankRequest request);
-}
