--- langchain4j-voyage/src/main/java/dev/langchain4j/model/voyage/VoyageJsonUtils.java
@@ -1,50 +1,0 @@
-package dev.langchain4j.model.voyage;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.SerializationFeature;
-
-class VoyageJsonUtils {
-
-    private VoyageJsonUtils() throws InstantiationException {
-        throw new InstantiationException("Can't instantiate this utility class.");
-    }
-
-    private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper()
-            .setPropertyNamingStrategy(PropertyNamingStrategies.SNAKE_CASE)
-            .enable(SerializationFeature.INDENT_OUTPUT)
-            .setSerializationInclusion(JsonInclude.Include.NON_NULL)
-            .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-
-    static String toJson(Object object) {
-        try {
-            return OBJECT_MAPPER.writeValueAsString(object);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    static <T> T toObject(String jsonStr, Class<T> clazz) {
-        try {
-            return OBJECT_MAPPER.readValue(jsonStr, clazz);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    static <T> T toObject(String jsonStr, TypeReference<T> typeReference) {
-        try {
-            return OBJECT_MAPPER.readValue(jsonStr, typeReference);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    static ObjectMapper getObjectMapper() {
-        return OBJECT_MAPPER;
-    }
-}
