+++ pom.xml
@@ -46,7 +46,7 @@
         <module>langchain4j-vertex-ai-gemini</module>
         <module>langchain4j-workers-ai</module>
         <module>langchain4j-zhipu-ai</module>
-        <module>langchain4j-voyage</module>
+        <module>langchain4j-voyage-ai</module>
 
         <!-- embedding stores -->
         <module>langchain4j-azure-ai-search</module>
